Update README.md
This program is designed to read in an unknown number of student ID numbers and scores then print both lists. They will then be sorted high to low based on test scores and printed once more. The list will then be sorted high to low based on ID number and printed. Then both lists will be sorted high to low based on test scores and printed. The lowest score, highest score, and the corresponding ID number will be printed along with an average for all test scores.